Add Nav component with title of current Kata
This is one solution to addressing the problem where we want to be able to know the name of the current Kata in order to know which file to edit.

Some of the Katas have `<Text />` labels with their names which works very well. However, for some Katas, the `<Text />` would interfere with the styling of the Kata itself.

This approach instead adds a Nav bar at the top of the Kata runner in the simulator. This allows us to both see the name of the current Kata as well as render the Kata without interference.

 * Install NavigationBar from react-native-navbar
 * Create own Nav component wrapping NavigationBar
 * Wrap runner in View with Nav and Kata as children